Rename shader_preprocessor_wgsl to shader_preprocessor
