+++ tfjs-backend-webgpu/src/kernels/matmul_small_output_size_webgpu.ts
@@ -16,7 +16,7 @@
  */
 
 import {backend_util, TensorInfo, util} from '@tensorflow/tfjs-core';
-import {getMainHeaderString} from '../shader_preprocessor_wgsl';
+import {getMainHeaderString} from '../shader_preprocessor';
 import {mapActivationToShaderProgram} from './activation_util';
 import {WebGPUProgram} from './webgpu_program';
